+++ pkg/ingester/ingester_v2.go
@@ -4,13 +4,10 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"math"
 	"net/http"
 	"os"
 	"path/filepath"
-	"sort"
-	"strconv"
 	"sync"
 	"time"
 
@@ -279,8 +276,6 @@ func (i *Ingester) startingV2(ctx context.Context) error {
 	}
 
 	// Scan and open backfill TSDB's that already exist on disk.
-	// TODO(codesome): run this in parallel with opening main TSDBs
-	// to make full use of TSDB opening concurrency.
 	if err := i.openExistingBackfillTSDB(context.Background()); err != nil {
 		return errors.Wrap(err, "opening existing backfill TSDBs")
 	}
@@ -472,8 +467,9 @@ func (i *Ingester) v2Push(ctx context.Context, req *client.WriteRequest) (*clien
 			}
 
 			cause := errors.Cause(err)
-			if cause == storage.ErrOutOfBounds && s.TimestampMs > db.Head().MaxTime()-i.cfg.TSDBConfig.BackfillLimit.Milliseconds() {
-				fmt.Println(s.TimestampMs, db.Head().MaxTime(), i.cfg.TSDBConfig.BackfillLimit.Milliseconds())
+			if cause == storage.ErrOutOfBounds &&
+				i.cfg.TSDBConfig.BackfillLimit > 0 &&
+				s.TimestampMs > db.Head().MaxTime()-i.cfg.TSDBConfig.BackfillLimit.Milliseconds() {
 				err := i.v2BackfillPush(userID, ts.Labels, s)
 				if err == nil {
 					succeededSamplesCount++
@@ -952,29 +948,45 @@ func (i *Ingester) getOrCreateTSDB(userID string, force bool) (*userTSDB, error)
 // createTSDB creates a TSDB for a given userID, and returns the created db.
 func (i *Ingester) createTSDB(userID string) (*userTSDB, error) {
 	tsdbPromReg := prometheus.NewRegistry()
-	udir := i.cfg.TSDBConfig.BlocksDir(userID)
-	userLogger := util.WithUserID(userID, util.Logger)
 
 	blockRanges := i.cfg.TSDBConfig.BlockRanges.ToMilliseconds()
+	userDB, err := i.createNewTSDB(
+		userID,
+		i.cfg.TSDBConfig.BlocksDir(userID),
+		blockRanges[0],
+		blockRanges[len(blockRanges)-1],
+		tsdbPromReg,
+	)
+	if err != nil {
+		// We set the limiter here because we don't want to limit
+		// series during WAL replay.
+		userDB.limiter = i.limiter
+		i.TSDBState.tsdbMetrics.setRegistryForUser(userID, tsdbPromReg)
+	}
+
+	return userDB, err
+}
+
+// createNewTSDB creates a TSDB for a given userID and data directory.
+func (i *Ingester) createNewTSDB(userID, dbDir string, minBlockDuration, maxBlockDuration int64, reg *prometheus.Registry) (*userTSDB, error) {
+	userLogger := util.WithUserID(userID, util.Logger)
 
 	userDB := &userTSDB{
 		userID:              userID,
 		refCache:            cortex_tsdb.NewRefCache(),
-		seriesInMetric:      newMetricCounter(i.limiter),
 		ingestedAPISamples:  newEWMARate(0.2, i.cfg.RateUpdatePeriod),
 		ingestedRuleSamples: newEWMARate(0.2, i.cfg.RateUpdatePeriod),
 		lastUpdate:          atomic.NewInt64(0),
 	}
 
 	// Create a new user database
-	db, err := tsdb.Open(udir, userLogger, tsdbPromReg, &tsdb.Options{
-		RetentionDuration:       i.cfg.TSDBConfig.Retention.Milliseconds(),
-		MinBlockDuration:        blockRanges[0],
-		MaxBlockDuration:        blockRanges[len(blockRanges)-1],
-		NoLockfile:              true,
-		StripeSize:              i.cfg.TSDBConfig.StripeSize,
-		WALCompression:          i.cfg.TSDBConfig.WALCompressionEnabled,
-		SeriesLifecycleCallback: userDB,
+	db, err := tsdb.Open(dbDir, userLogger, reg, &tsdb.Options{
+		RetentionDuration: i.cfg.TSDBConfig.Retention.Milliseconds(),
+		MinBlockDuration:  minBlockDuration,
+		MaxBlockDuration:  maxBlockDuration,
+		NoLockfile:        true,
+		StripeSize:        i.cfg.TSDBConfig.StripeSize,
+		WALCompression:    i.cfg.TSDBConfig.WALCompressionEnabled,
 	})
 	if err != nil {
 		return nil, errors.Wrapf(err, "failed to open TSDB: %s", udir)
@@ -991,9 +1003,6 @@ func (i *Ingester) createTSDB(userID string) (*userTSDB, error) {
 	}
 
 	userDB.DB = db
-	// We set the limiter here because we don't want to limit
-	// series during WAL replay.
-	userDB.limiter = i.limiter
 	userDB.setLastUpdate(time.Now()) // After WAL replay.
 
 	// Thanos shipper requires at least 1 external label to be set. For this reason,
@@ -1013,8 +1022,8 @@ func (i *Ingester) createTSDB(userID string) (*userTSDB, error) {
 	if i.cfg.TSDBConfig.ShipInterval > 0 {
 		userDB.shipper = shipper.New(
 			userLogger,
-			tsdbPromReg,
-			udir,
+			reg,
+			dbDir,
 			cortex_tsdb.NewUserBucketClient(userID, i.TSDBState.bucket),
 			func() labels.Labels { return l },
 			metadata.ReceiveSource,
@@ -1022,7 +1031,6 @@ func (i *Ingester) createTSDB(userID string) (*userTSDB, error) {
 		)
 	}
 
-	i.TSDBState.tsdbMetrics.setRegistryForUser(userID, tsdbPromReg)
 	return userDB, nil
 }
 
@@ -1381,436 +1389,4 @@ func (i *Ingester) v2FlushHandler(w http.ResponseWriter, _ *http.Request) {
 	}()
 
 	w.WriteHeader(http.StatusNoContent)
-}
-
-// createBackfillTSDB creates a TSDB for a given userID and bucketName, and returns the created db.
-func (i *Ingester) createBackfillTSDB(userID, bucketName string, blockRange int64) (*userTSDB, error) {
-	tsdbPromReg := prometheus.NewRegistry()
-	udir := i.cfg.TSDBConfig.BackfillBlocksDir(userID)
-	dbDir := filepath.Join(udir, bucketName)
-	userLogger := util.WithUserID(userID, util.Logger)
-
-	userDB := &userTSDB{
-		userID:              userID,
-		refCache:            cortex_tsdb.NewRefCache(),
-		ingestedAPISamples:  newEWMARate(0.2, i.cfg.RateUpdatePeriod),
-		ingestedRuleSamples: newEWMARate(0.2, i.cfg.RateUpdatePeriod),
-		lastUpdate:          atomic.NewInt64(0),
-	}
-
-	// Create a new user database
-	db, err := tsdb.Open(dbDir, userLogger, tsdbPromReg, &tsdb.Options{
-		RetentionDuration: i.cfg.TSDBConfig.Retention.Milliseconds(),
-		MinBlockDuration:  blockRange,
-		MaxBlockDuration:  blockRange,
-		NoLockfile:        true,
-		StripeSize:        i.cfg.TSDBConfig.StripeSize,
-		WALCompression:    i.cfg.TSDBConfig.WALCompressionEnabled,
-	})
-	if err != nil {
-		return nil, err
-	}
-	db.DisableCompactions() // we will compact on our own schedule
-
-	userDB.DB = db
-	userDB.setLastUpdate(time.Now()) // After WAL replay.
-
-	// Thanos shipper requires at least 1 external label to be set. For this reason,
-	// we set the tenant ID as external label and we'll filter it out when reading
-	// the series from the storage.
-	l := labels.Labels{
-		{
-			Name:  cortex_tsdb.TenantIDExternalLabel,
-			Value: userID,
-		}, {
-			Name:  cortex_tsdb.IngesterIDExternalLabel,
-			Value: i.lifecycler.ID,
-		},
-	}
-
-	// Create a new shipper for this database
-	if i.cfg.TSDBConfig.ShipInterval > 0 {
-		userDB.shipper = shipper.New(
-			userLogger,
-			tsdbPromReg,
-			udir,
-			cortex_tsdb.NewUserBucketClient(userID, i.TSDBState.bucket),
-			func() labels.Labels { return l },
-			metadata.ReceiveSource,
-			true, // Allow out of order uploads. It's fine in Cortex's context.
-		)
-	}
-
-	//i.TSDBState.tsdbMetrics.setRegistryForUser(userID, tsdbPromReg)
-	return userDB, nil
-}
-
-func (i *Ingester) v2BackfillPush(userID string, la []client.LabelAdapter, s client.Sample) error {
-	bucket, err := i.getOrCreateBackfillTSDB(userID, s.TimestampMs)
-	if err != nil {
-		return err
-	}
-
-	startAppend := time.Now()
-	db := bucket.db
-	bucket.lastAppend = startAppend
-	cachedRef, cachedRefExists := db.refCache.Ref(startAppend, client.FromLabelAdaptersToLabels(la))
-
-	app := db.Appender()
-	// If the cached reference exists, we try to use it.
-	if cachedRefExists {
-		err = app.AddFast(cachedRef, s.TimestampMs, s.Value)
-		if err != nil && errors.Cause(err) == storage.ErrNotFound {
-			cachedRefExists = false
-			err = nil
-		}
-	}
-
-	// If the cached reference doesn't exist, we (re)try without using the reference.
-	if !cachedRefExists {
-		var ref uint64
-
-		// Copy the label set because both TSDB and the cache may retain it.
-		copiedLabels := client.FromLabelAdaptersToLabelsWithCopy(la)
-
-		if ref, err = app.Add(copiedLabels, s.TimestampMs, s.Value); err == nil {
-			db.refCache.SetRef(startAppend, copiedLabels, ref)
-			cachedRef = ref
-			cachedRefExists = true
-		}
-	}
-
-	if err == nil {
-		err = app.Commit()
-	}
-
-	return err
-}
-
-// Assumes 1h bucket range for . TODO(codesome): protect stuff with locks.
-type backfillTSDBs struct {
-	userStateMtx sync.RWMutex
-	tsdbs        map[string][]*tsdbBucket
-}
-
-func newBackfillTSDBs() *backfillTSDBs {
-	return &backfillTSDBs{
-		tsdbs: make(map[string][]*tsdbBucket),
-	}
-}
-
-func (b *backfillTSDBs) getBucketsForUser(userID string) []*tsdbBucket {
-	b.userStateMtx.RLock()
-	defer b.userStateMtx.RUnlock()
-	return b.tsdbs[userID]
-}
-
-func (i *Ingester) getOrCreateBackfillTSDB(userID string, ts int64) (*tsdbBucket, error) {
-	userBuckets := i.TSDBState.backfillDBs.getBucketsForUser(userID)
-
-	start, end := getBucketRangesForTimestamp(ts, 1)
-
-	var bucket *tsdbBucket
-	insertIdx := len(userBuckets)
-	for idx, b := range userBuckets {
-		if model.Time(ts) >= b.bucketStart && model.Time(ts) < b.bucketEnd {
-			bucket = b
-			break
-		}
-
-		// Existing: |-----------|
-		//      New:       |------------|
-		if b.bucketStart < start && start < b.bucketEnd {
-			start = b.bucketEnd
-		}
-
-		// Existing:         |-----------|
-		//      New:  |------------|
-		if end > b.bucketStart && end < b.bucketEnd {
-			end = b.bucketStart
-			insertIdx = idx
-			break
-		}
-
-		if b.bucketStart > end {
-			insertIdx = idx
-			break
-		}
-	}
-
-	if bucket == nil {
-		tsdb, err := i.createBackfillTSDB(userID, getBucketName(start, end), int64(end-start)*2)
-		if err != nil {
-			return nil, err
-		}
-		bucket = &tsdbBucket{
-			db:          tsdb,
-			lastAppend:  time.Now(),
-			bucketStart: start,
-			bucketEnd:   end,
-		}
-		userBuckets = append(userBuckets[:insertIdx], append([]*tsdbBucket{bucket}, userBuckets[insertIdx:]...)...)
-		i.TSDBState.backfillDBs.userStateMtx.Lock()
-		i.TSDBState.backfillDBs.tsdbs[userID] = userBuckets
-		i.TSDBState.backfillDBs.userStateMtx.Unlock()
-	}
-
-	return bucket, nil
-}
-
-func (i *Ingester) openExistingBackfillTSDB(ctx context.Context) error {
-	level.Info(util.Logger).Log("msg", "opening existing TSDBs")
-	wg := &sync.WaitGroup{}
-	openGate := gate.New(i.cfg.TSDBConfig.MaxTSDBOpeningConcurrencyOnStartup)
-
-	users, err := ioutil.ReadDir(i.cfg.TSDBConfig.BackfillDir)
-	if err != nil {
-		return err
-	}
-
-	var runErr error
-	for _, u := range users {
-		if !u.IsDir() {
-			continue
-		}
-
-		userID := u.Name()
-		userPath := filepath.Join(i.cfg.TSDBConfig.BackfillDir, userID)
-
-		bucketNames, err := ioutil.ReadDir(userPath)
-		if err != nil {
-			level.Error(util.Logger).Log("msg", "unable to open user TSDB dir for backfill", "err", err, "user", u, "path", userPath)
-			continue
-		}
-
-		for _, bucketName := range bucketNames {
-			if bucketName.IsDir() {
-				continue
-			}
-
-			dbPath := filepath.Join(userPath, bucketName.Name())
-			f, err := os.Open(dbPath)
-			if err != nil {
-				level.Error(util.Logger).Log("msg", "unable to open user backfill TSDB dir", "err", err, "user", userID, "path", dbPath)
-				return filepath.SkipDir
-			}
-			defer f.Close()
-
-			// If the dir is empty skip it
-			if _, err := f.Readdirnames(1); err != nil {
-				if err != io.EOF {
-					level.Error(util.Logger).Log("msg", "unable to read backfill TSDB dir", "err", err, "user", userID, "path", dbPath)
-				}
-
-				return filepath.SkipDir
-			}
-
-			// Limit the number of TSDB's opening concurrently. Start blocks until there's a free spot available or the context is cancelled.
-			if err := openGate.Start(ctx); err != nil {
-				runErr = err
-				break
-			}
-
-			wg.Add(1)
-			go func(userID, bucketName string) {
-				defer wg.Done()
-				defer openGate.Done()
-				defer func(ts time.Time) {
-					i.TSDBState.walReplayTime.Observe(time.Since(ts).Seconds())
-				}(time.Now())
-
-				start, end, err := getBucketRangesForBucketName(bucketName)
-				db, err := i.createBackfillTSDB(userID, bucketName, int64(end-start)*2)
-				if err != nil {
-					level.Error(util.Logger).Log("msg", "unable to open user backfill TSDB", "err", err, "user", userID)
-					return
-				}
-
-				bucket := &tsdbBucket{
-					db:          db,
-					lastAppend:  time.Now(),
-					bucketStart: start,
-					bucketEnd:   end,
-				}
-
-				i.TSDBState.backfillDBs.userStateMtx.Lock()
-				// Append at the end, we will sort it at the end.
-				i.TSDBState.backfillDBs.tsdbs[userID] = append(i.TSDBState.backfillDBs.tsdbs[userID], bucket)
-				i.TSDBState.backfillDBs.userStateMtx.Unlock()
-			}(userID, bucketName.Name())
-		}
-
-		if runErr != nil {
-			break
-		}
-
-	}
-
-	// Wait for all opening routines to finish
-	wg.Wait()
-
-	// Sort the buckets within the users.
-	i.TSDBState.backfillDBs.userStateMtx.Lock()
-	for _, buckets := range i.TSDBState.backfillDBs.tsdbs {
-		sort.Slice(buckets, func(i, j int) bool {
-			return buckets[i].bucketStart < buckets[i].bucketEnd
-		})
-	}
-	i.TSDBState.backfillDBs.userStateMtx.Unlock()
-
-	return runErr
-}
-
-func (i *Ingester) backfillSelect(ctx context.Context, userID string, from, through int64, matchers []*labels.Matcher) ([]storage.SeriesSet, error) {
-	buckets := i.TSDBState.backfillDBs.getBucketsForUser(userID)
-
-	var queriers []storage.Querier
-	defer func() {
-		for _, q := range queriers {
-			q.Close()
-		}
-	}()
-	for _, b := range buckets {
-		if !b.overlaps(from, through) {
-			mint, err := b.db.DB.StartTime()
-			if err != nil {
-				return nil, err
-			}
-			maxt := b.db.Head().MaxTime()
-			if !overlapsOpenInterval(mint, maxt, from, through) {
-				continue
-			}
-		}
-
-		q, err := b.db.Querier(ctx, from, through)
-		if err != nil {
-			return nil, err
-		}
-
-		queriers = append(queriers, q)
-	}
-
-	if len(queriers) == 0 {
-		return nil, nil
-	}
-
-	result := make([]storage.SeriesSet, len(queriers))
-	errC := make(chan error, 1)
-	var wg sync.WaitGroup
-	for i, q := range queriers {
-		wg.Add(1)
-		go func(i int, q storage.Querier) {
-			defer wg.Done()
-
-			ss := q.Select(false, nil, matchers...)
-			if ss.Err() != nil {
-				select {
-				case errC <- ss.Err():
-				default:
-				}
-			}
-			result[i] = ss
-		}(i, q)
-	}
-
-	wg.Wait()
-	select {
-	case err := <-errC:
-		return nil, err
-	}
-
-	return result, nil
-}
-
-type tsdbBucket struct {
-	db                     *userTSDB
-	bucketStart, bucketEnd model.Time
-	lastAppend             time.Time
-}
-
-func (b *tsdbBucket) overlaps(mint, maxt int64) bool {
-	return overlapsOpenInterval(int64(b.bucketStart), int64(b.bucketEnd), mint, maxt)
-}
-
-func overlapsOpenInterval(mint1, maxt1, mint2, maxt2 int64) bool {
-	return mint1 < maxt2 && mint2 < maxt1
-}
-
-// getBucketName returns the string representation of the bucket.
-// YYYY_MM_DD_HH_YYYY_MM_DD_HH
-func getBucketName(start, end model.Time) string {
-	startTime := start.Time()
-	endTime := end.Time()
-
-	return fmt.Sprintf(
-		"%04d_%02d_%02d_%02d_%04d_%02d_%02d_%02d",
-		startTime.Year(), startTime.Month(), startTime.Day(), startTime.Hour(),
-		endTime.Year(), endTime.Month(), endTime.Day(), endTime.Hour(),
-	)
-}
-
-func getBucketRangesForTimestamp(ts int64, bucketSize int) (model.Time, model.Time) {
-	// TODO(codesome): Replace this entire thing with 1-2 simple equations
-	// to align ts with the nearest hours which also aligns with bucketSize.
-	t := time.Unix(ts/1000, 0).UTC()
-	yyyy := t.Year()
-	mm := t.Month()
-	dd := t.Day()
-	hh := bucketSize * (t.Hour() / bucketSize)
-	t = time.Date(yyyy, mm, dd, hh, 0, 0, 0, time.UTC)
-
-	start := model.Time(t.Unix() * 1000)
-	end := start.Add(time.Duration(bucketSize) * time.Hour)
-
-	return start, end
-}
-
-func getBucketRangesForBucketName(bucketName string) (model.Time, model.Time, error) {
-	// YYYY_MM_DD_HH_YYYY_MM_DD_HH
-	// 012345678901234567890123456
-	if len(bucketName) != 27 {
-		return 0, 0, errors.New("Invalid bucket name")
-	}
-
-	startYYYY, err := strconv.Atoi(bucketName[0:4])
-	if err != nil {
-		return 0, 0, err
-	}
-	startMM, err := strconv.Atoi(bucketName[5:7])
-	if err != nil {
-		return 0, 0, err
-	}
-	startDD, err := strconv.Atoi(bucketName[8:10])
-	if err != nil {
-		return 0, 0, err
-	}
-	startHH, err := strconv.Atoi(bucketName[11:13])
-	if err != nil {
-		return 0, 0, err
-	}
-
-	endYYYY, err := strconv.Atoi(bucketName[14:18])
-	if err != nil {
-		return 0, 0, err
-	}
-	endMM, err := strconv.Atoi(bucketName[19:21])
-	if err != nil {
-		return 0, 0, err
-	}
-	endDD, err := strconv.Atoi(bucketName[22:24])
-	if err != nil {
-		return 0, 0, err
-	}
-	endHH, err := strconv.Atoi(bucketName[25:27])
-	if err != nil {
-		return 0, 0, err
-	}
-
-	startTime := time.Date(startYYYY, time.Month(startMM), startDD, startHH, 0, 0, 0, time.UTC)
-	endTime := time.Date(endYYYY, time.Month(endMM), endDD, endHH, 0, 0, 0, time.UTC)
-
-	start := model.Time(startTime.Unix() * 1000)
-	end := model.Time(endTime.Unix() * 1000)
-
-	return start, end, nil
 }
